Rely on security group ID when deleting it
As OpenStack allows resources to have the same name, it's
possible that we end up with multiple security groups with
same name, and the deletion of those resources for UPI fails
as currently we're relying on sg Name. This commit fixes the
issue by ensuring the resoruce ID is used instead.