Bug:1842871, oVirt - template temporary VM is created with no tag
At the beginning of the installation the installer creates a temporary VM which it uses to create a template.
The temorary VM and the bootstrap VM are created with no tag, we want the bootstrap VM and the temporary VM
to be destroied on destroy bootstrap.
This commit adds a ovirt tag "${ocp_cluster_id}-bootstrap" to the temporary VM and bootstrap VM which will
be picked up on destroy bootstrap.

Signed-off-by: Gal-Zaidman <gzaidman@redhat.com>