GCP: Fix uninstaller to not use hardcoded base path.
The uninstaller currently hard codes the base path as the GCP
API base path has changed.

Change 1:
For the getZoneName function, the InstanceGroupAggregatedList,
InstanceAggregatedList and the disk list functions that calls
this to get the proper url all have a shortened URL as the key
in the list.Items map. Example of a shortened URL is
zones/us-central1-a.
Reusing the key value instead of hard coding the base path
and TrimLeft operation.

Change 2:
For the getInstanceGroupURL function, the call to the underlying
InstanceGroupAggregatedList to get the InstanceGroups already have
the URL to hit called the SelfLink which is now set to the returning
object and used for the URL, thus not requiring the call.

Change 3:
For the getInstanceNameAndZone, since the underlying object has only
the URL and no other relevant information, the base path is removed
and the TrimLeft operation is changed to Split based on the '/projects/'
URL query parameter instead of using the base path.

Previous commit information:
https://github.com/openshift/installer/pull/2745/commits/11c9773466bdcd034e8de3068f31bd140298515c#diff-4781d154886c4da10a007bb219f27d8aL170
TODO: Find a better solution for change 3.