create cmd: Display command asset directories.
Added a feature where the directories of the assets that were
used during any target of the create command is invoked, are
displayed on the screen. Examples of the display that will show
up are

'''
INFO Please find the install-config in the test directory
INFO Please find the manifests in the test/openshift and test/manifests directories
INFO Please find the ignition-configs in the test/auth and test directories
INFO Please find the cluster in the test/auth, test/tls and test directories
'''

If a file is created that is 2 or more folders deep, the first folder name is displayed.