Add the --pod-infra-container-image flag to the kubelet service
Override the --pod-infra-container-image flag to point to the pause
image from the release payload. By default, this flag is set to k8s.gcr.io/pause:3.1
and the kubelet asks the runtime for the image status of this image every few minutes,
which is why we were seeing the warning logs in cri-o saying that this image was not
found. This is avoided by making the pod-infra-container-image flag point to the actual
pause image being used from the release payload.

Signed-off-by: Urvashi Mohnani <umohnani@redhat.com>