azure/machines.go: make sure virtual machines are added to public LB as soon as they are created
On Azure the route to internet depends on various factors but for openshift where all machines have internal ips,
whether machine belongs to a public lb backend decides how the traffic is routed to the internet.

when the machine boots up, it is not part of any LB backend so all the traffic flows based on scenario 3 in [1],
while as soon as k8s cloud provider adds the node to the public LB backend the traffic flows based on scenario 2 in [1].
To make sure that traffic to internet takes one path, it's useful to add the machine to the LB backed as soon as it is created. Since there
is only one public LB since 7c1a274c , we can use the machine-api to add the machines to correct backend as soon as they are created.

For outbound type UDR, the machines do not have to be added to the LB as even when machine is added to the LB by k8s cloud provider, it does
not affect the egress as it is controlled by however the user has setup the outbound routing.

[1]: https://docs.microsoft.com/en-us/azure/load-balancer/load-balancer-outbound-connections