baremetal: bump ironic timeout to 3600 seconds
terraform-provider-ironic recently changed the approach it uses for
timeouts from an estimate based on number of attempts, to using a
go routine in a context. The logic it used to estimate was off by more
than a factor of two. That means the current 1500 second timeout was
actually more than 3000.

This commit restores the previously unintended behavior as now
explicitly set. The latest RHCOS images are around 800 MB in size. At
256kB/s, this would be 3,125 seconds to download. We also have to allow
time for the bootstrap VM itself to boot, which we could alot 5 minutes
for, totaling 3,425 seconds. I round it up to 3600 just for the even
number.

This timeout is unfortunate but I think is required for users who
download the images directly from the internet and not a local cache.
Ideally we would be able to monitor in real time what the
machine-os-downloader container is doing, and fail earlier if an error
it encountered was fatal.