Bump terraform-provider-openstack
Version 1.28.0 changes the default authentication against OpenStack to
lazily request tokens and reuse them.

In order to minimise the impact of this and the next upgrade, this
change bumps from 1.25 to 1.28 at the beginning of the OpenShift release
cycle.

The new Terraform provider in turns requires a bump in Gophercloud.