baremetal: vendor: bump baremetal-operator
This updates baremetal-operator to pull in the latest changes including
hardware profiles.  The installer vendors hardware profiles from
baremetal-operator to keep the interface the same between
terraform-provisioned baremetal hosts and those provisioned by the
machine-api.

This is done by updating the entry in the `replace` clause of the go.mod
file.  Then, we run `go mod tidy`, followed by `go mod vendor`. This
breaks because the requested version of google.golang.org/grpc, v1.27.0,
changes the API in backwards-incompatible ways.  Thus, we also include a
new pin for this dependency to the previously requested version,
v1.25.0, which yields a working build.