gcp: Fix uninstaller for internal clusters on GCP.
Fixed the GCP uninstaller to get the instance group self links which
would be used to destroy them. Currently, the self links are used but
are not requested from GCP as a field when the instance groups are listed.
Therefore, the uninstaller does not get the appropriate link to target
for deletion.