baremetal: make provisioning network optional
This implements the enhancement request that makes the provisioning
network optional. It introduces a new platform configuration called
"provisioningNetwork" that takes 3 possible values: managed, unmanaged,
and disabled. The "unmanaged" option replaces the external DHCP flag
that was previously used.

When the network is set to "disabled," it is expected that users use
virtualmedia configuration. We will deploy no DHCP/TFTP provisioning
services for a user to use.