machines/gcp: use GCP image when licenses are not required
The machines will use the osImage as is for re-using the RHCOS published images. But for
cases where the licenses are set for platform, the installer will have to create a new GCP image
using the raw image. Therefore set the image for machines when platform.Licenses is provided to `<cluster-id>-rhcos-image`
to match the name of the image that will be created later on.