gcp: only create new image when licenses are specified
Tf templates are provided with `gcp_image_uri` which is the location of the raw image, and
`gcp_image` which is the gcp image sources from one of the master machine objects.

Tf templates always use the `gcp_image` which is a pre-existing image published by RHCOS, but
in cases like where the Licenses are set for the GCP platform we will create a new image using the uri.