baremetal: Prevent race condition when adding HardwareDetails
We add HardwareDetails to a BareMetalHost resource by including them in
initial Status data that is attached to the Host as an annotation. This
annotation must be present before the BareMetalHost is first reconciled
(thus creating the Status subresource), otherwise it will be ignored.

It is not possible to add the annotation at the time the BareMetalHost
is created. Therefore, we were previously relying on it being added
before the baremetal-operator is up and running. In practice this works
because the baremetal-operator takes a long time to start up, but in
fact this is a race.

Prevent the race by setting the baremetalhost.metal3.io/paused
annotation on the CR at the time it is created, and removing it again
when the status annotation is added. This annotation prevents the
operator from attempting to reconcile the resource.