Remove the need for externally built libvirt CI images.
Currently, libvirt testing requires images built outside of CI, and not maintained by the installer team.
These images are built manually with packer out of https://github.com/ironcladlou/openshift4-libvirt-gcp/blob/rhel8/IMAGES.md
Instead, the provision script that is run when creating these external images should be maintained here,
and kept in line with any changes or updates in the installer repository.

Also baked into the libvirt CI images is a script that runs the install.  This also should live here and
be maintained by the installer team.

There will be a transition period in the libvirt test environment while these scripts are moved from external
images to the libvirt-installer image.  Once the transition is complete, the libvirt-installer image
no longer has to build the installer (saving time in the CI tests).  Instead, the installer will be
extracted from the provided release image (in CI, it's `${RELEASE_IMAGE_LATEST}`).