OpenStack: Deleting servers using tag-based filtering
Now to delete servers we first get a list of all available servers
from Nova, and then we iterate through them to find those with
required metadata. In the case of a large number of servers, this
can take a very long time.

Fortunately gophercloud introduced filtering by tags, so we can
start using this feature to get only servers with the required tag.
https://github.com/gophercloud/gophercloud/pull/1759