vsphere: Add validations for IP inputs
With the current setup, it is possible for the user to enter the
same IP address for the ingress and the API for vsphere. Creating
a cluster with these settings results in Bootstrap and one of the
 masters being assigned the same IP address during installation.

Added a validation to check if the given input for the Ingress is
not the same as API whose input would have been specified first
and throw an error if they are the same, prompting the user to
re-enter the IP address again.