Bug 1852112: Ignore hostPrefix validation for non-(sdn/ovn) plugins
For plugins like Calico which don't use the hostPrefix for anything,
it does not make sense to validate this field and force customers
to set specific values just to pass validation. These plugins should
be able to unset the hostPrefix completely and not have to worry about
it. This patch ensures that we perform the validation only for
OpenShiftSDN and OVNKubernetes. In the case of other plugins, we
check if the hostPrefix is unset(0) and if so, we skip the validation.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>