baremetal: Enable installer survey for baremetal platform
This PR enables installer survey for the baremetal platform where the user is asked a minimal set of questions, which combined with sane defaults to construct an installer-config file. The config file can either directly used to create a cluster or can be customized by editing it manually before creating the cluster.