destroy: means to obtain gcp blocked resources
Add a `GetAllPendingItems` function to the GCP destroyer that the
user can use to get all of the resources that were not able to be
destroyed.

This will be used by Hive to expose to the user the resources that
cannot be deleted so that the user can take action on those resources.

https://issues.redhat.com/browse/CO-974