OpenStack: Create SG rules in parallel
Terraform-provider-openstack suffered from a race condition that would
lead to conflicts in neutron and forced us to create security group
sequentially (via an explicit dependency on the previous rule) to
prevent failures. This is no longer the case in
terraform-provider-openstack v1.28.0 where this issue was fixed with
[1].

With the recent bump of terraform-provider-openstack to v1.28.0, we can
now remove the workaround.

This reverts commit 228aadf9ebde1b60490bff19511de49d6350008f.

[1] https://github.com/terraform-providers/terraform-provider-openstack/pull/994