Bug 1857188: OpenStack: skip container deletion if it was removed
During deletion of containers, we get a list of available containers
first, and then we iterate through them to find the ones we need,
based on their metadata.
Since this is not an atomic operation, it may happen that the containers
can be removed at runtime. We should ignore these cases and continue
to iterate through the remaining ones.