asset/cluster: tag the user provided rg with owned tag
The cluster resource group when created by the installer is tagged with `kubernetes.io_cluster_<infraID>=owned` tag.
This tag is used to clean up the service principles generated by cred-minter and also to mark that the resource group is owned by
this specific cluster.

Since user provided resource group still is owned by the cluster, we must tag this resource group similarly. Since terraform cannot
update the tags for an existing resource group, we have to fall back the pattern of `PreTerraform` previously used by AWS for pre-existing subnets.

We read existing tags from the resource group and add `kubernetes.io_cluster_<infraID>=owned` to it when sending an update for the resource group.