baremetal: create machines and machineset linked to cluster
The installer takes the cluster name given by the user and appends a
random string. MachineSets and Machines create within the cluster need
to be linked to that cluster in order to be recognized by some of the
controllers in the machine API. This is especially important, for
example, when scaling a MachineSet to ensure that the new Machines
created by the machine-set-controller are linked to the MachineSet
properly. See https://bugzilla.redhat.com/show_bug.cgi?id=1857175 for
more details.

This change drops the use of cluster name in the MachineSet and
Machine resoures created for a baremetal cluster and uses the
clusterID instead. The baremetal platform was the only one using name
instead of ID in this way.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>