ovirt: Create cloud provider config
Motivation
Some cluster components can use the platform details to create more
capabilities. One example is CSI driver, where the we want to initialize
a default storage class, with the StorageDomainID as one of its
parameters. The benefit would be the ability to generate a default
storage class with the details of the cloud provider config map.

Result
A config map with oVirt's cloud provider config

Notes
oVirt does not have a supported external cloud provider or a cloud
provider config yet. What we have is this [1] and it may be used for future
reference.

[1] https://github.com/oVirt/ovirt-openshift-extensions/blob/master/cmd/ovirt-cloud-provider/ovirt-cloud-provider.go

Signed-off-by: Roy Golan <rgolan@redhat.com>