oVirt: remove tmp template VM on bootstrap phase
This patch breaks the dependency between terraform resources:
releaseimage_template and tmp_import_vm

On ovirt before the bootstrap, masters and workers start they need to be created from a template.
The installer creates this template with the terraform template module.
As part of the process for creating the template, we create an oVirt tmp VM which should be sealed into a template for the bootstrap, masters and workers machines and then removed.
Until this PR the tmp VM will just stay in the oVirt cluster until cluster destroy is called, this is not good because it is just garbage in the oVirt cluster that the user will probably remove on its own once the installation is complete.
This PR makes bootstrap destroy command to destroy the tmp VM as well, which make sense because it is created for bootstrapping the cluster.

Signed-off-by: Gal-Zaidman <gzaidman@redhat.com>