baremetal: handle boot mode for hosts
Give the user a bootMode field for each host so they can control whether
the host should be booted using UEFI or BIOS mode. Pass the value to
Ironic for control plane nodes and include it in the BareMetalHost
resources created for all hosts.

Change the Ironic parameter to use instance_info/deploy_boot_mode instead
of the capabilities, based on feedback from the Ironic team.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>