Validations and Doc Updates for Floating IPs
The user interface for floating IPs has been changed such that
the ingress and api floating ips can no longer be passed when
the external network is unset. These validations check that this
usage is respected, while also checking the the floating IPs passed
by the user are valid and available.