data/aws: add variable for ignition stub for bootstrap
Currently the terraform code creates a stub ignition that points to s3 bucket. This restricts us
to depend on ignition provider to create an appropiate stub in cases like addition trust for cloud APIs.

Therefore accepting the stub as is allows the installer to create a stub ignition and pass it own as-is for terraform.