installer commands: Modify installer to send metrics
This adds the calls in the installer to the metrics gatherer APIs
to create, pass information and send metrics to Prometheus
Aggregation Gateway.

Added utility functions that the installer can use to do common
functions like setting the success messages and return code, log
errors, send data etc. Modified timer code to return a duration
object which would be used to set the histogram values. Added code
to create the three types of metrics - invocation, duration and
modification metrics.