vsphere upi: Configure haproxy to check /readyz
Current hapropy setting for the vsphere upi is to check the
6443 port to see if the api server is up in tcp mode. TCP breaks
graceful termination and /readyz provides error information before
the api server crashes which would give some time for the load
balancer to react.

Adding a check in the HAProxy server to check for the /readyz
port for 200 status.

The health check matches the one in ipi deployments for apiserver lb hosted in the
cluster. see [1]

[1]: https://github.com/openshift/machine-config-operator/blob/2ca16131e4bd0a2c967196e5b4e3961a829335f8/templates/master/00-master/vsphere/files/vsphere-haproxy-haproxy.yaml#L38-L44