Bug 1866534: fixes bootstrap DNS failure on bare metal
The fix is also applied to openstack, ovirt and vsphere platforms.

The local-dns-prepender ensures that during installation, the bootstrap host is
able to resolve the cluster's own domain and subdomains using its own DNS
server running on localhost. It does this by ensuring that localhost is the
first entry in resolv.conf, and it is triggered any time an interface is
activated.

In case an interface gets a new dhcp4 or dhcp6 lease that updates DNS entries,
the local-dns-prepender will not be activated, and its changes to resolv.conf
will be overwritten.

This change adds the "dhcp4-change" and "dhcp6-change" action types to the
local-dns-prepender for bare metal. In testing, it resolved the issue I saw
where bootkube.sh would never complete due to inability to resolve the API
FQDN for the new control plane.

fixes #3988