cmd: gather logs on cluster asset creation failure
If we fail to create the cluster asset (i.e. run terraform), no logs get
gathered because we exit fatally. This changes the create target to try
to gather logs before exiting. It does not run the entire PostRun
function.

This is useful on some platforms where terraform may fail but the
bootstrap host may be up enough to capture logs.