baremetal: rename JSON field for ClusterProvisioningIP
The ClusterProvisioningIP field in the baremetal platform had an
incorrect JSON annotation for serialization/deserialization, using
"provisioningHostIP" instead of "clusterProvisioningIP." This name is
incorrect and doesn't adequately reflect it's purpose. This field is
the IP used in the cluster for provisioning.

This commit deprecates the old JSON field, and uses the correct JSON
annotation for ClusterProvisioningIP.