Libvirt: Bump bootstrap memory to 4G
Noticed on 4.6 installs that processes on the bootstrap node were getting OOM killed. This was mainly due to
large memory consumption on by the kube-apiserver (https://bugzilla.redhat.com/show_bug.cgi?id=1868773) although
this looks to be normal as other platforms have a bigger memory limit for the bootstrap node.