Set proxy.config/cluster spec.trustedCA if have CA
An install config may specify an additional trusted bundle without
specifying any proxy configuration.  Before this commit, such an install
config would produce a "user-ca-bundle" configmap but would not configure
anything to use it.  After this commit, such an install config produces the
"user-ca-bundle" configmap and also adds a reference to the configmap in
the spec.trustedCA field of the proxy.config/cluster object.  As a result,
cluster-network-operator will read "user-ca-bundle" and inject the
additional trusted bundle into the "trusted-ca-bundle" configmap, which
other operators, such as the authentication operator, use.

This commit is related to bug 1866818.

https://bugzilla.redhat.com/show_bug.cgi?id=1866818

* pkg/asset/manifests/proxy.go (Generate): Set spec.trustedCA if the
install config specifies an additional trusted bundle, irrespective of
whether proxy configuration is specified.