Embed rhcos*.json as /bootimage/rhcos*.json in container image
This is intended as a short term tweak on the (long) road
to https://github.com/openshift/enhancements/pull/201

See specifically this subthread:
https://github.com/openshift/enhancements/pull/201#discussion_r375546207
And https://github.com/openshift/enhancements/pull/201#issuecomment-665183467

TL;DR we want `oc` to be the frontend to this data, via something like
`oc adm release info --bootimage-json` or so, and not `openshift-install`.

We want to move towards having the data as part of the release image.
But the installer will have RHCOS pinned for the near future because
of the bootstrap node - we need an RHCOS for that too.

And it's going to be very problematic to have "two sources of truth"
for RHCOS even temporarily.

This small change puts the RHCOS json inside the `installer` container
image which is already part of the release payload and built from
this repository; then we can patch `oc` to have `oc adm bootimages --json`
which would do the equivalent of:
`oc image extract /bootimage:. $(oc adm release info --image-for=installer))`