WIP: docs/user/vsphere: Replace static hostname/IP to use Afterburn
I didn't test this myself, but we should have landed the work
for 4.6.  This is much more ergonomic for users than appending
to the Ignition configs, and also solves the core problem that
we need networking set up in the initramfs to fetch the real
config from the Machine Config Server.