oVirt: remove tmp VM disk on destroy
When calling terraform destroy the tmpVM disk is not removed from the engine. This is because the tmp VM unlike any other VM is not created from an oVirt template, and we create a separate disk for it with ovirt_image_transfer.
This PR removes the disk resource when calling destroy

Signed-off-by: Gal-Zaidman <gzaidman@redhat.com>