Add support to clean custom router
To support a fip less installation and bring your
own network when using Kuryr, new interfaces are
added to a custom Router that may exist, enabling
traffic between Pods, Services and VMs. As the
Router is not created by the installer its interfaces
must be cleaned up upon cluster Destroy.

This commit solves the issue by discovering the
Router through the Primary Network and the gateway
interface attached to the Router.