asset/quota/gcp: use GCP api to find CPU count for constraint and guess only on failure
Previously we were using the machine type name to figure out the CPU count and metric name. But it seems like using the GCP API
will be more future proof compared to string matching.

So now the constraint calculator will fetch the CPU count from GCP api and only on failure guess the count using string matching.
Fallback to the string matching allows us to guess the CPU count when API access to not available or other failures. The guess always
drops the count to 0 when it cannot guess with high confidence.

Also adds unit tests for the machine type -> guess cpu function and use GCP api and fallback to guess function.