machines/gcp: add support for labeling resources
GCP supports labeling resources [1] to help group them together in a
lightweight way. These labels should not be confused with tags, which
are a different but related concept.

This change adds support for user-specified labels that should be
applied to GCP resources. Similar functionality in the installer is
already available for other cloud providers (such as AWS).

[1] https://cloud.google.com/compute/docs/labeling-resources