Fix cluster destroy on a FIP less installation
In case the machines Subnet is not connected to a Router
there is not a need to clean-up the interfaces from the
custom Router as no additional interfaces would have
been created on it. Also, when using Kuryr if a Service
of LoadBalancer type was created, a floating ip would get
created for the load balancer and the removal of the service
subnet from the router would be blocked.

This commit fixes both issues by ignoring the custom Router
clean-up when no Router is found and moving the custom
Router clean-up to after the load balancers removal.