openstack UPI: Allow for no FIP or router
If the external network is not provided, the router will not be created
and external connectivity will be a respopnsibility of the operator.

Floating IPs will be set if available, on an individual basis. The
external network is a requirement for FIPs; if not provided, the
playbooks will error.

The bootstrap FIP is now required to be created manually like the other
FIPs.