Bug 1880896: do not use local cacert path in generated clouds.yaml
When we generate an OpenShift cred secret with clouds.yaml, we keep the
local cacert path there. In other words, path to cacert on the user's
machine. This is wrong, because this value is useless after the
installation. To avoid this we need to use the predefined const value,
like we do with cloud config:
https://github.com/openshift/installer/blob/4aac205a59a9102ddf1ae329b3c4c6f0dff7e330/pkg/asset/manifests/openstack/cloudproviderconfig.go#L21