Bug 1777224: pkg/asset: metadata to depend on ignition
Prior to this change, the Metadata asset was only regenerated when the
install-config or install-config.infraID was determined to have changed.
However, in the case that the install-config was updated, the installer
was invoked with `create manifests`, and then the installer was invoked
a second time with `create ignition-configs`, the second invokation had
no way of knowing the install-config had changed. As a result, the
Metadata would not be regenerated. This caused the clusterID in the
metadata.json to not match that of the cluster.

This change solves this specific case by making the Metadata asset
dependent on the bootstrap ignition asset. It in turn is dependent on
the manifests asset. As a result, when the manifests are regenerated
in the first invocation, and the bootstrap ignition is regenerated in
the second invocation, the Metadata will also be regenerated.