Bug 1882649: enable image import again
In 4.4 we disabled image import mechanism because it was not stable
enough on the testing environment. Now we expect that it works better,
so we can enable image import uploading by default.
But, to prevent possible failures, the installer falls back to legacy
uploading if image import fails to upload data.