baremetal: Don't complete until master-bmh-update is done
We are not done with the bootstrap host until the master-bmh-update
script has successfully completed, adding the IP address data needed by
the master nodes, and enabling reconciliation on the master Host
objects. To ensure the bootstrap cannot go away before this is done,
delay signalling that bootstrapping is complete until it is finished.