[Azure][Destroy] Check if resource group exists
Azure clubs all resources for a cluster within a resource group. The
pattern for delete is to first delete all the resources and then delete
the resource group.

However, when attempting delete, we should first check if the
resource group exists and then try to delete the resources. This is to
ensure delete exits gracefully when resource group wasn't created and/or
deleted before attempting delete.