cmd/openshift-install/create: Replace Routes().List with Get
We know the name we're looking for, so a Get should be slightly more
efficient.  The list landed with the route logic in ff53523b11 (add
logs at end of install for kubeadmin, consoleURL, 2018-12-06, #806)
without clear motivation.  It's possible that the intention was
shielding from the console operator renaming their route, with the:

  Route found in openshift-console namespace: ...

lines pointing us at the renamed route.  But it seems unlikely that
any such renaming would happen, and if it does, we expect the
console-operator change to fail its end-to-end testing with the
installer not finding a route with the old name.